Add HTTP routes and image cache

+++ utils/utils.go
@@ -3,6 +3,6 @@ package utils
 import "image/color"
 
 type Pixel struct {
-	X, Y   int
+	X, Y   int64
 	Colour color.RGBA
 }
